Better control over frequency and bladeRF/LMS.
Added a PIO component to the Qsys so the NIOS core can reset the LMS as
well as toggle the RX/TX enable pins, and switch the TX and RX paths
between low band and high band operation.

Modified frequency selection algorithm for fractional-N VCO which should
now be accurate within a few Hz.